+++ scripts/generate_docs.py
@@ -82,6 +82,7 @@ CATAGORYS = {
         "security_role_data_source.md",
         "security_roles_data_source.md",
         "security_role_resource.md",
+        "security_login_message_resource.md",
     ],
     'snaplock': [],
     'snapmirror': [